+++ src/main/kotlin/in/kkkev/jjidea/jj/JujutsuRepository.kt
@@ -33,7 +33,7 @@ data class JujutsuRepositoryImpl(override val project: Project, override val dir
     JujutsuRepository {
     private val executor: CommandExecutor by lazy {
         val settings = JujutsuSettings.getInstance(project)
-        CliExecutor(directory, settings.state.jjExecutablePath)
+        CliExecutor(directory) { settings.state.jjExecutablePath }
     }
 
     /**